refactor(frontend): isolate gpu replica field
